requirements/base.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
pillow==10.2.0
psycopg2==2.9.9